mk802ii: Disable unused soc parts
The .fex file as found in the android boot partition is a copy and paste
job from previous EVB (evaluation board) .fex files, as the name tag in
there clearly indicates. But the mk802ii does not come with an ethernet
socket, lcd, tv-out, etc. So lets make the fex file match the actual hardware
and disable those parts of the A10-soc.

This fixes for example an unusable eth0 interface showing up in "ifconfig -a"

Signed-off-by: Hans de Goede <hdegoede@redhat.com>