mini-x-1gb: Disable unused soc parts
The .fex file as found in the android boot partition is a copy and paste
job from previous EVB (evaluation board) .fex files, as the name tag in
there clearly indicates. But the mini-x-1gb does not use the 2nd and 3th
i2c controllers, the 2nd uart, the lcd, etc. So lets make the fex file match
the actual hardware and disable those parts of the A10-soc.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>