uhost_u1a: Enable usb2-controller at boot
Like on most boards the original fex has usb_host_init_state set to 0 for
usbc2, causing it to start with its USB-port power powered down.

This causes any attached devices (the realtek wireless card to be precise)
to not get detected, and the kernel-module to not be auto-loaded.

Manually modprobing the driver works around this, as it causes a bus scan
which powers up the port. But manual modprobing should not be necessary, usb
is a plug and play bus afterall. Setting usb_host_init_state to 1 removes the
need for the manual modprobe, now Linux will automatically find the device
and udev will automatically load the driver.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>