uhost_u1a, mk802+_a10s: Add playback_used = 0 to audio_para
playback_used is a linux-sunxi.org fex file extension for boards which use
the onboard audio codec, but only for input, such as the uhost_u1a, and the
A10s based mk802. Yes the A10s based mk802 has a microphone on the board,
there is no holer over it in the cover, so I wonder how well it works, but
it is there.

Note that since this is a linux-sunxi.org extension the kernel code will
assume "playback_used = 1" when it is not present in the fex file.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>