mele_a3700: Add new fex file for the mele a3700
I noticed there is a Mele_A3700 board type in u-boot, but no fex file
in sunxi-boards. Looking at: http://linux-sunxi.org/Mele_A3700

And specifically at:
http://linux-sunxi.org/images/1/1a/Mele_A3700_PCB_TOPa.jpg

I've noticed that it uses the same PCB as my a1000g, without
a sata connecter soldered onto the board. So this patch adds a new
mele_a3700.fex which is a copy of mele_a1000g.fex, with the sata
controller disabled.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>