various: Correct iowidth for models which have 4 8bit RAM chips
livesuit generates wrong dram settings for these devices, so the iowidth
we've extracted from the Android image settings is wrong.

Note there may be other cases of the same issue, but this patch only
changes those boards of which I'm sure they have 4 8bit RAM chips in all
their incarnations.

This is not always the case! IE for the mini-x-1gb there are both versions
with 2 - 16 * 256 Mbit RAM chips, as well as versions with 4 - 8 * 256 Mbit
RAM chips, in which case it is safest to use an iowidth of 16.

Note this has a matching commit in u-boot sunxi-current, since the actual
dram settings are done by u-boot, the dram part of the fex-file is just
informational.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>