various: Enable pull-up on uart pins
Many boards have the uart pins hooked up to test points, or jumper pins for
debugging use, so most of the time they end up floating.

Sometimes this results in reception of spurious messages. To fix this
allwinner has enabled the pull ups for the uart pins in its evb-1.2 fex
file.

Since most of our fex files are based on the evb-1.2 fex, they already
have the pull ups enabled. But some are based on the older evb-1.1 fex,
this patch enables the pull ups there too.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>