A33: A-Star KV49L tablet
Add an initial fex for the KV49L. This fex was obtained by dumping the
ram via /dev/mem, with the advantage that the ram para's seem to have
been filled in properly. For some reason however, some sections came out
to '0' and bin2fex didn't know how to handle it, so I left it as is for
now.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>