a13: Add fex file for the Empire Electronix D709 tablet
The Empire Electronix D709 tablet is a fairly standard 7" A13 tablet,
featuring usb-wifi, a micro-sd slot, micro-usb otg and headphone jack.

Empire Electronix is written on the bakc of the tablet, the D709 model
info can be found in the about tablet menu in android.

The PCB has no markings to speak of.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>