Add fex file from my two A33 tablets
One is Aoson M751S, a Q8-compatible tablet, with a 800x480 TFT LCD,
GSL1680 touch controller, MMA8653 accelerometer and ESP8089 WLAN card.

Another is a unbranded with a PCB number of iNet_D978_Rev2, with a
1024x768 IPS LCD, Goodix GT9271 touch controller, unknown accelerometer
and RTL8723BS WLAN/BT combo card, and with a touch home key (come with a
LED). Onda's V975s Quad v7 is like this one, but with a RTL8818EU
rather than RTL8723BS. (Other versions of V975s come with A31s or A80)

Signed-off-by: Icenowy Zheng <icenowy@aosc.xyz>